requirements.txt
fastapi==0.111.0
pydantic==2.7.4
starlette==0.37.2
uvicorn==0.29.0
python-dotenv==1.0.1
requests==2.31.0
putergenai==0.1.5
slowapi==0.1.9